Cleaner pytest skip condition for sharding tests.
